Ends map loading/unloading for game selection from UI
