+++ bonus/srcs_bonus/animation_bonus/destroy_animation_bonus.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   destroy_animation_bonus.c                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: brappo <brappo@student.42.fr>              +#+  +:+       +#+        */
+/*   By: mhotting <mhotting@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/24 12:32:35 by brappo            #+#    #+#             */
-/*   Updated: 2024/08/27 13:41:11 by brappo           ###   ########.fr       */
+/*   Updated: 2024/09/05 14:08:49 by mhotting         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
