+++ bonus/srcs_bonus/game_bonus/game_level_selection1_bonus.c
@@ -6,7 +6,7 @@
 /*   By: mhotting <mhotting@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/09/03 14:47:03 by mhotting          #+#    #+#             */
-/*   Updated: 2024/09/05 09:21:28 by mhotting         ###   ########.fr       */
+/*   Updated: 2024/09/05 14:31:28 by mhotting         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,5 +16,13 @@ void	game_reach_level_selection(t_game *game)
 {
 	if (game == NULL)
 		return ;
+	game->state = STATE_LEVEL_SELECTION;
+}
+
+void	quit_level(t_game *game)
+{
+	if (game == NULL)
+		return ;
+	game_unload_map(game);
 	game->state = STATE_LEVEL_SELECTION;
 }
