+++ bonus/srcs_bonus/game_bonus/t_game_init_bonus.c
@@ -16,6 +16,52 @@
 #include "event_handlers_bonus.h"
 #include "mlx.h"
 
+static bool	init_textures(
+	t_game *game,
+	t_map *map,
+	t_animation anim[MAP_NB_IDS]
+)
+{
+	size_t			index;
+	t_mlx_coords	*texture_size;
+	char			*filename;
+	t_image			texture;
+
+	index = 0;
+	while (index < MAP_NB_IDS)
+	{
+		filename = map->textures[index];
+		texture_size = &map->texture_size[index];
+		if (filename == NULL || filename[0] == '\0')
+			return (error_print_string(ERR_MISSING_TEXTURES, filename), false);
+		if (!check_extension(filename, ".xpm"))
+			return (error_print_string(ERR_TEXTURE_EXTENSION, filename), false);
+		if (!t_image_import_file(&texture, filename, game->mlx.mlx_ptr, NULL))
+			return (error_print_string(ERR_INIT_TEXTURES, filename), false);
+		anim[index].textures = create_animation_textures(&texture, \
+			texture_size, game->mlx.mlx_ptr, filename);
+		mlx_destroy_image(game->mlx.mlx_ptr, texture.ptr);
+		if (anim[index].textures == NULL)
+			return (false);
+		index++;
+	}
+	return (true);
+}
+
+static bool	init_all_map_textures(t_game *game)
+{
+	size_t	i;
+
+	i = 0;
+	while (i < NB_MAPS)
+	{
+		if (!init_textures(game, &game->maps[i], game->anims[i]))
+			return (false);
+		i++;
+	}
+	return (true);
+}
+
 static bool	init_draw_thread_args(
 	t_game *game,
 	t_draw_thread_arg draw_thread_args[NB_DRAW_THREADS]
@@ -71,5 +117,6 @@ bool	t_game_init(t_game *game)
 		&& init_game_uis(game, &game->uis)
 		&& t_mlx_launch(&game->mlx)
 		&& add_event_handlers(game)
+		&& init_all_map_textures(game)
 	);
 }
